+++ lib/src/dtos/record_model.dart
@@ -2,6 +2,7 @@ import "dart:convert";
 
 import "package:json_annotation/json_annotation.dart";
 
+import "../caster.dart" as caster;
 import "jsonable.dart";
 
 part "record_model.g.dart";
@@ -75,159 +76,50 @@ class RecordModel implements Jsonable {
     return model;
   }
 
-  /// Returns a single [data] value with key [fieldName] as **String**.
+  /// Extracts a single value from [data] by a dot-notation path
+  /// and tries to cast it to the specified generic type.
   ///
-  /// For `null` values empty string is returned.
-  /// `toString()` is used for any other non-String value.
-  String getStringValue(String fieldName) {
-    final rawValue = data[fieldName];
-
-    if (rawValue == null) {
-      return "";
-    }
-
-    return rawValue is String ? rawValue : rawValue.toString();
+  /// If explicitly set, returns [defaultValue] on missing path.
+  ///
+  /// For more details about the casting rules, please refer to
+  /// [caster.extract()].
+  ///
+  /// Example:
+  ///
+  /// ```dart
+  /// final data = {"a": {"b": [{"b1": 1}, {"b2": 2}, {"b3": 3}]}};
+  /// final record = RecordModel(data: data);
+  /// final result0 = record.getDataValue<int>("a.b.c", "missing"); // "missing"
+  /// final result1 = record.getDataValue<int>("a.b.2.b3"); // 3
+  /// final result2 = record.getDataValue<String>("a.b.2.b3"); // "3"
+  /// ```
+  T getDataValue<T>(String fieldNameOrPath, [T? defaultValue]) {
+    return caster.extract<T>(data, fieldNameOrPath, defaultValue);
   }
 
-  /// Returns a single [data] value with key [fieldName] as List.
-  ///
-  /// Non-List fields will be wrapped (eg. `1` will be returned as `[1]`).
-  List<T> getListValue<T>(String fieldName) {
-    final rawValue = data[fieldName];
-
-    if (rawValue is List) {
-      return rawValue.cast<T>();
-    }
-
-    if (rawValue is T) {
-      return <T>[rawValue];
-    }
-
-    return <T>[];
+  /// An alias for [getDataValue<String>()].
+  String getStringValue(String fieldNameOrPath, [String? defaultValue]) {
+    return getDataValue<String>(fieldNameOrPath, defaultValue);
   }
 
-  /// Returns a single [data] value with key [fieldName] as **bool**.
-  ///
-  /// For non-num values the following casting rules are applied:
-  /// - `null` - always returned as `false`
-  /// - `num` - `false` if `0`, otherwise `true`
-  /// - `String` - `false` if one of `"", "false", "0"`, otherwise - `true`
-  /// - `List` and `Map` - `true` if `length > 0`, otherwise - `false`
-  /// - `false` for any other type
-  bool getBoolValue(String fieldName) {
-    final rawValue = data[fieldName];
-
-    if (rawValue == null) {
-      return false;
-    }
-
-    if (rawValue is bool) {
-      return rawValue;
-    }
-
-    if (rawValue is num) {
-      return rawValue != 0;
-    }
-
-    if (rawValue is String) {
-      final falsyValues = <String>["", "false", "0"];
-
-      return !falsyValues.contains(rawValue.toLowerCase());
-    }
-
-    if (rawValue is Iterable) {
-      return rawValue.isNotEmpty;
-    }
-
-    if (rawValue is Map) {
-      return rawValue.isNotEmpty;
-    }
-
-    return false;
+  /// An alias for [getDataValue<List<T>>()].
+  List<T> getListValue<T>(String fieldNameOrPath, [List<T>? defaultValue]) {
+    return getDataValue<List<T>>(fieldNameOrPath, defaultValue);
   }
 
-  /// Returns a single [data] value with key [fieldName] as **int**.
-  ///
-  /// For non-num values the following casting rules are applied:
-  /// - `null` - always returned as `0`
-  /// - `String` - the non-null result of `int.tryParse()`, otherwise -`0`.
-  /// - `bool` - `false` -> `0`, `true` -> `1`
-  /// - `List` and `Map` - returns the length of the List/Map.
-  /// - `0` for any other type
-  int getIntValue(String fieldName) {
-    final rawValue = data[fieldName];
-
-    if (rawValue == null) {
-      return 0;
-    }
-
-    if (rawValue is int) {
-      return rawValue;
-    }
-
-    if (rawValue is double) {
-      return rawValue.toInt();
-    }
-
-    if (rawValue is String) {
-      return int.tryParse(rawValue) ?? 0;
-    }
-
-    if (rawValue is bool) {
-      return rawValue ? 1 : 0;
-    }
-
-    if (rawValue is Iterable) {
-      return rawValue.length;
-    }
-
-    if (rawValue is Map) {
-      return rawValue.length;
-    }
-
-    return 0;
+  /// An alias for [getDataValue<bool>()].
+  bool getBoolValue(String fieldNameOrPath, [bool? defaultValue]) {
+    return getDataValue<bool>(fieldNameOrPath, defaultValue);
   }
 
-  /// Returns a single [data] value with key [fieldName] as **double**.
-  ///
-  /// For non-num values the following casting rules are applied:
-  /// - `null` - always returned as `0`
-  /// - `String` - the non-null result of `double.tryParse()`, otherwise -`0`.
-  /// - `bool` - `false` -> `0`, `true` -> `1`
-  /// - `List` and `Map` - returns the length of the List/Map.
-  /// - `0` for any other type
-  double getDoubleValue(String fieldName) {
-    final rawValue = data[fieldName];
+  /// An alias for [getDataValue<int>()].
+  int getIntValue(String fieldNameOrPath, [int? defaultValue]) {
+    return getDataValue<int>(fieldNameOrPath, defaultValue);
+  }
 
-    if (rawValue == null) {
-      return 0;
-    }
-
-    if (rawValue is double) {
-      return rawValue;
-    }
-
-    if (rawValue is int) {
-      return rawValue.toDouble();
-    }
-
-    if (rawValue is String) {
-      return double.tryParse(rawValue) ?? 0;
-    }
-
-    if (rawValue is bool) {
-      return rawValue ? 1 : 0;
-    }
-
-    if (rawValue is Iterable) {
-      return rawValue.length.toDouble();
-    }
-
-    if (rawValue is Map) {
-      return rawValue.length.toDouble();
-    }
-
-    return 0;
+  /// An alias for [getDataValue<double>()].
+  double getDoubleValue(String fieldNameOrPath, [double? defaultValue]) {
+    return getDataValue<double>(fieldNameOrPath, defaultValue);
   }
 
   @override