+++ test/dtos/record_model_test.dart
@@ -47,132 +47,153 @@ void main() {
       expect(model.toJson(), json);
     });
 
-    //test("getStringValue()", () {
-    //  final model = RecordModel(data: {
-    //    "a": null,
-    //    "b": 1,
-    //    "c": "test",
-    //    "d": false,
-    //    "e": ["1", 2, 3],
-    //    "f": {"test": 123},
-    //  });
-    //  expect(model.getStringValue("a"), "");
-    //  expect(model.getStringValue("b"), "1");
-    //  expect(model.getStringValue("c"), "test");
-    //  expect(model.getStringValue("d"), "false");
-    //  expect(model.getStringValue("e"), "[1, 2, 3]");
-    //  expect(model.getStringValue("f"), "{test: 123}");
-    //});
+    test("getDataValue()", () {
+      final model = RecordModel(data: {
+        "a": null,
+        "b": 1.5,
+        "c": "test",
+        "d": false,
+        "e": ["1", "2", "3"],
+        "f": {"test": 123},
+      });
 
-    //test("getListValue()", () {
-    //  final model = RecordModel(data: {
-    //    "a": null,
-    //    "b": 1,
-    //    "c": "test",
-    //    "d": false,
-    //    "e": ["1", 2, 3],
-    //  });
-    //  expect(model.getListValue<String>("a"), <String>[]); // invalid type
-    //  expect(model.getListValue<dynamic>("a"), [null]);
-    //  expect(model.getListValue<bool>("b"), <bool>[]); // invalid type
-    //  expect(model.getListValue<num>("b"), <num>[1]); // invalid type
-    //  expect(model.getListValue<dynamic>("b"), [1]);
-    //  expect(model.getListValue<dynamic>("c"), ["test"]);
-    //  expect(model.getListValue<dynamic>("d"), [false]);
-    //  expect(model.getListValue<dynamic>("e"), ["1", 2, 3]);
-    //});
+      expect(model.getDataValue<dynamic>("unknown"), null);
+      expect(model.getDataValue("unknown", "missing!"), "missing!");
+      expect(model.getDataValue<num>("b"), 1.5);
+      expect(model.getDataValue<int>("b"), 1);
+      expect(model.getDataValue<String>("c"), "test");
+      expect(model.getDataValue<List<String>>("e"), ["1", "2", "3"]);
+      expect(model.getDataValue<Map<String, dynamic>>("f"), {"test": 123});
+      expect(model.getDataValue<int>("f.test"), 123);
+      expect(model.getDataValue<int>("f.test2", -1), -1);
+    });
 
-    //test("getBoolValue()", () {
-    //  final model = RecordModel(data: {
-    //    "a": null,
-    //    "b1": 1,
-    //    "b2": 0,
-    //    "b3": -1,
-    //    "c1": "",
-    //    "c2": "false",
-    //    "c3": "0",
-    //    "c4": "test",
-    //    "d": false,
-    //    "e1": <dynamic>[],
-    //    "e2": ["1", 2, 3],
-    //    "f1": <String, dynamic>{},
-    //    "f2": {"test": 123},
-    //  });
-    //  expect(model.getBoolValue("a"), false);
-    //  expect(model.getBoolValue("b1"), true);
-    //  expect(model.getBoolValue("b2"), false);
-    //  expect(model.getBoolValue("b3"), true);
-    //  expect(model.getBoolValue("c1"), false);
-    //  expect(model.getBoolValue("c2"), false);
-    //  expect(model.getBoolValue("c3"), false);
-    //  expect(model.getBoolValue("c4"), true);
-    //  expect(model.getBoolValue("d"), false);
-    //  expect(model.getBoolValue("e1"), false);
-    //  expect(model.getBoolValue("e2"), true);
-    //  expect(model.getBoolValue("f1"), false);
-    //  expect(model.getBoolValue("f2"), true);
-    //});
+    test("getStringValue()", () {
+      final model = RecordModel(data: {
+        "a": null,
+        "b": 1,
+        "c": "test",
+        "d": false,
+        "e": ["1", 2, 3],
+        "f": {"test": 123},
+      });
+      expect(model.getStringValue("a"), "");
+      expect(model.getStringValue("b"), "1");
+      expect(model.getStringValue("c"), "test");
+      expect(model.getStringValue("d"), "false");
+      expect(model.getStringValue("e"), "[1, 2, 3]");
+      expect(model.getStringValue("f"), "{test: 123}");
+    });
 
-    //test("getIntValue()", () {
-    //  final model = RecordModel(data: {
-    //    "a": null,
-    //    "b1": 1,
-    //    "b2": 2.4,
-    //    "c1": "",
-    //    "c2": "false",
-    //    "c3": "123",
-    //    "c4": "test",
-    //    "d1": false,
-    //    "d2": true,
-    //    "e1": <dynamic>[],
-    //    "e2": ["1", 2, 3],
-    //    "f1": <String, dynamic>{},
-    //    "f2": {"a": 123, "b": 456},
-    //  });
-    //  expect(model.getIntValue("a"), 0);
-    //  expect(model.getIntValue("b1"), 1);
-    //  expect(model.getIntValue("b2"), 2);
-    //  expect(model.getIntValue("c1"), 0);
-    //  expect(model.getIntValue("c2"), 0);
-    //  expect(model.getIntValue("c3"), 123);
-    //  expect(model.getIntValue("c4"), 0);
-    //  expect(model.getIntValue("d1"), 0);
-    //  expect(model.getIntValue("d2"), 1);
-    //  expect(model.getIntValue("e1"), 0);
-    //  expect(model.getIntValue("e2"), 3);
-    //  expect(model.getIntValue("f1"), 0);
-    //  expect(model.getIntValue("f2"), 2);
-    //});
+    test("getListValue()", () {
+      final model = RecordModel(data: {
+        "a": null,
+        "b": 1,
+        "c": "test",
+        "d": false,
+        "e": ["1", 2, 3],
+      });
+      expect(model.getListValue<String>("a"), <String>[]); // invalid type
+      expect(model.getListValue<dynamic>("a"), [null]);
+      expect(model.getListValue<bool>("b"), <bool>[]); // invalid type
+      expect(model.getListValue<num>("b"), <num>[1]); // invalid type
+      expect(model.getListValue<dynamic>("b"), [1]);
+      expect(model.getListValue<dynamic>("c"), ["test"]);
+      expect(model.getListValue<dynamic>("d"), [false]);
+      expect(model.getListValue<dynamic>("e"), ["1", 2, 3]);
+    });
 
-    //test("getDoubleValue()", () {
-    //  final model = RecordModel(data: {
-    //    "a": null,
-    //    "b1": 1,
-    //    "b2": 2.4,
-    //    "c1": "",
-    //    "c2": "false",
-    //    "c3": "123.4",
-    //    "c4": "test",
-    //    "d1": false,
-    //    "d2": true,
-    //    "e1": <dynamic>[],
-    //    "e2": ["1", 2, 3],
-    //    "f1": <String, dynamic>{},
-    //    "f2": {"a": 123, "b": 456},
-    //  });
-    //  expect(model.getDoubleValue("a"), 0);
-    //  expect(model.getDoubleValue("b1"), 1);
-    //  expect(model.getDoubleValue("b2"), 2.4);
-    //  expect(model.getDoubleValue("c1"), 0);
-    //  expect(model.getDoubleValue("c2"), 0);
-    //  expect(model.getDoubleValue("c3"), 123.4);
-    //  expect(model.getDoubleValue("c4"), 0);
-    //  expect(model.getDoubleValue("d1"), 0);
-    //  expect(model.getDoubleValue("d2"), 1);
-    //  expect(model.getDoubleValue("e1"), 0);
-    //  expect(model.getDoubleValue("e2"), 3);
-    //  expect(model.getDoubleValue("f1"), 0);
-    //  expect(model.getDoubleValue("f2"), 2);
-    //});
+    test("getBoolValue()", () {
+      final model = RecordModel(data: {
+        "a": null,
+        "b1": 1,
+        "b2": 0,
+        "b3": -1,
+        "c1": "",
+        "c2": "false",
+        "c3": "0",
+        "c4": "test",
+        "d": false,
+        "e1": <dynamic>[],
+        "e2": ["1", 2, 3],
+        "f1": <String, dynamic>{},
+        "f2": {"test": 123},
+      });
+      expect(model.getBoolValue("a"), false);
+      expect(model.getBoolValue("b1"), true);
+      expect(model.getBoolValue("b2"), false);
+      expect(model.getBoolValue("b3"), true);
+      expect(model.getBoolValue("c1"), false);
+      expect(model.getBoolValue("c2"), false);
+      expect(model.getBoolValue("c3"), false);
+      expect(model.getBoolValue("c4"), true);
+      expect(model.getBoolValue("d"), false);
+      expect(model.getBoolValue("e1"), false);
+      expect(model.getBoolValue("e2"), true);
+      expect(model.getBoolValue("f1"), false);
+      expect(model.getBoolValue("f2"), true);
+    });
+
+    test("getIntValue()", () {
+      final model = RecordModel(data: {
+        "a": null,
+        "b1": 1,
+        "b2": 2.4,
+        "c1": "",
+        "c2": "false",
+        "c3": "123",
+        "c4": "test",
+        "d1": false,
+        "d2": true,
+        "e1": <dynamic>[],
+        "e2": ["1", 2, 3],
+        "f1": <String, dynamic>{},
+        "f2": {"a": 123, "b": 456},
+      });
+      expect(model.getIntValue("a"), 0);
+      expect(model.getIntValue("b1"), 1);
+      expect(model.getIntValue("b2"), 2);
+      expect(model.getIntValue("c1"), 0);
+      expect(model.getIntValue("c2"), 0);
+      expect(model.getIntValue("c3"), 123);
+      expect(model.getIntValue("c4"), 0);
+      expect(model.getIntValue("d1"), 0);
+      expect(model.getIntValue("d2"), 1);
+      expect(model.getIntValue("e1"), 0);
+      expect(model.getIntValue("e2"), 3);
+      expect(model.getIntValue("f1"), 0);
+      expect(model.getIntValue("f2"), 2);
+    });
+
+    test("getDoubleValue()", () {
+      final model = RecordModel(data: {
+        "a": null,
+        "b1": 1,
+        "b2": 2.4,
+        "c1": "",
+        "c2": "false",
+        "c3": "123.4",
+        "c4": "test",
+        "d1": false,
+        "d2": true,
+        "e1": <dynamic>[],
+        "e2": ["1", 2, 3],
+        "f1": <String, dynamic>{},
+        "f2": {"a": 123, "b": 456},
+      });
+      expect(model.getDoubleValue("a"), 0);
+      expect(model.getDoubleValue("b1"), 1);
+      expect(model.getDoubleValue("b2"), 2.4);
+      expect(model.getDoubleValue("c1"), 0);
+      expect(model.getDoubleValue("c2"), 0);
+      expect(model.getDoubleValue("c3"), 123.4);
+      expect(model.getDoubleValue("c4"), 0);
+      expect(model.getDoubleValue("d1"), 0);
+      expect(model.getDoubleValue("d2"), 1);
+      expect(model.getDoubleValue("e1"), 0);
+      expect(model.getDoubleValue("e2"), 3);
+      expect(model.getDoubleValue("f1"), 0);
+      expect(model.getDoubleValue("f2"), 2);
+    });
   });
 }